+++ src/TenderBody/Tender.java
@@ -1,6 +1,10 @@
 package TenderBody;
 
-import java.util.Objects;
+import java.util.*;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+import Exception.UnsuitableArgumentsException;
 
 public class Tender {
 
@@ -12,31 +16,69 @@ public class Tender {
         this.quantity = quantity;
     }
 
-    public Tender() {
+    public Tender(){
     }
 
     public Skills getSkills() {
         return skills;
     }
 
-    public void setSkills(Skills skills) {
-        this.skills = skills;
-    }
-
     public int getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
-        this.quantity = quantity;
+    public Map<String, Integer> convertTenderToMap(List<Tender> tender) {
+        return tender.stream().collect(Collectors.toMap(k -> String.valueOf(k.getSkills()), v -> v.getQuantity()));
     }
 
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Tender)) return false;
-        Tender tender = (Tender) o;
-        return getQuantity() == tender.getQuantity() && getSkills() == tender.getSkills();
+    public Map<String, Integer> convertWorkerToMap(List<Worker> workers) {
+        ArrayList sb = new ArrayList();
+        for (Worker w : workers) {
+            sb.add(w.getSkills());
+        }
+        String[] brigade = sb.toString().replaceAll("\\[|\\]|\\<|\\>", "")
+                .replaceAll("\\s+", "").split(",");
+
+        Map<String, Integer> mapWorkers = new TreeMap<>();
+        for (String worker : brigade) {
+            if (mapWorkers.keySet().contains(worker)) {
+                mapWorkers.put(worker, mapWorkers.get(worker) + 1);
+            } else {
+                mapWorkers.put(worker, 1);
+            }
+        }
+        return mapWorkers;
+    }
+
+    public boolean equalsBrigadeWithTender(List<Tender> tender, List<Worker> workers) {
+        return convertTenderToMap(tender)
+                .entrySet().containsAll(convertWorkerToMap(workers)
+                        .entrySet()) && convertWorkerToMap(workers)
+                .entrySet().containsAll(convertTenderToMap(tender).entrySet());
+    }
+
+    public List<Brigade> choiceBrigade(List<Tender> tender, List<Brigade> brigades) throws UnsuitableArgumentsException {
+        Integer minPrice = 0;
+        int countOfSuitableBrigades = 0;
+        List<Integer> prices = new ArrayList<>();
+        for (Brigade b : brigades) {
+            if (equalsBrigadeWithTender(tender, b.getWorkers())) {
+                prices.add(b.getPrice());
+                countOfSuitableBrigades++;
+            }
+        }
+        if (countOfSuitableBrigades == 0) {
+            throw new UnsuitableArgumentsException("There isn't a suitable brigade");
+        }
+        minPrice = Collections.min(prices);
+        System.out.println("Winner of tender with price = " + minPrice + ":");
+        List<Brigade> winner = new ArrayList<>();
+        for (Brigade brigade : brigades) {
+            if (brigade.getPrice() == minPrice) {
+                winner.add(brigade);
+            }
+        }
+        return winner;
     }
 
     @Override
